# packages/contracts/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AEP BFF API
  version: 1.0.0
servers:
- url: /api/v1
paths:
...
  /projects/{projectName}/files/{path}:
    get:
      operationId: read-file
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: 'File path under specs/ (trailing wildcard: may span multiple segments, i.e. contain
          slashes)'
        in: path
        name: path
        required: true
        schema:
          description: File path under specs/
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/FileContent
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
      security:
      - userJWT: []
      summary: Read a spec file at HEAD
      tags:
      - Files
  /projects/{projectName}/issues:
    get:
      operationId: list-issues
      summary: List/search GitHub issues on a project's repo
      tags:
      - Issues
      security:
      - userJWT: []
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          type: string
      - description: Comma-separated GitHub labels to filter by
        in: query
        name: labels
        schema:
          type: string
      - description: Keyword search over issue title/body, ranked by distinct-term
          overlap (title weighted double), capped at 25. Recall-biased — used to
          surface related issues before filing a new one.
        in: query
        name: q
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: ./components.yaml#/components/schemas/IssueInfo
                type: array
                nullable: true
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
    post:
      operationId: create-issue
      summary: Create a GitHub issue on a project's repo
      tags:
      - Issues
      security:
      - userJWT: []
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./components.yaml#/components/schemas/CreateIssueRequest
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/IssueResult
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
...
  /projects/{projectName}/tasks:
    get:
      operationId: list-tasks
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: Which Tasks to return (default open)
        explode: false
        in: query
        name: state
        schema:
          description: Which Tasks to return (default open)
          enum:
          - open
          - closed
          - all
          type: string
      - description: Filter to Tasks planned from this spec/build version tag (e.g. v3); empty returns
          all versions.
        explode: false
        in: query
        name: tag
        schema:
          description: Filter to Tasks planned from this spec/build version tag (e.g. v3); empty returns
            all versions.
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: ./components.yaml#/components/schemas/TaskView
                type: array
                nullable: true
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
      security:
      - userJWT: []
      summary: List a project's Tasks (live GitHub ⋈ executions → derived status), optionally scoped to
        one spec/build version tag
      tags:
      - Tasks
  /projects/{projectName}/tasks/{issueNumber}:
    get:
      operationId: get-task
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: GitHub issue number of the Task
        in: path
        name: issueNumber
        required: true
        schema:
          description: GitHub issue number of the Task
          format: int64
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/TaskDetail
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
      security:
      - userJWT: []
      summary: Get one Task with its full Execution history
      tags:
      - Tasks
...
  /projects/{projectName}/tasks/{issueNumber}/promote-from-issue:
    post:
      operationId: promote-task-from-issue
      summary: Turn an ad-hoc GitHub issue into a coding Task and dispatch it (async)
      tags:
      - Tasks
      security:
      - userJWT: []
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          type: string
      - description: GitHub issue number to promote and dispatch
        in: path
        name: issueNumber
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./components.yaml#/components/schemas/PromoteFromIssueRequest
        required: true
      responses:
        '202':
          description: Accepted — the funnel dispatches out-of-band.
        default:
          content:
            application/json:
              schema:
                $ref: ./components.yaml#/components/schemas/Error
          description: Error
...
components:
  securitySchemes:
    userJWT:
      bearerFormat: JWT
      description: Thunder end-user JWT (RS256, JWKS-verified)
      scheme: bearer
      type: http